backend/requirements.txt
asgiref==3.11.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
PyJWT==2.10.1
PyMySQL==1.1.0
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.5
wheel==0.45.1
